Fixed incorrect skill amounts being written to journal
The issue appears to stem from copying over the similar if-block from the unBlock function. Instead of multiplying like in unBlock, the code in catchJavaAddXP uses division, and the numbers were not modified to account for that. The variable xpBoostNumerator was also being used as the divisor, so I renamed it appropriately to 'denominator'. Left some robust comments to help potential readers understand the math as well.